pyproject.toml
[build-system]
requires = ["hatchling>=1.8.0", "hatch-vcs"]
build-backend = "hatchling.build"

[tool.hatch.version]
path = "src/dtscalibration/__init__.py"

[tool.hatch.build.targets.sdist]
exclude = [
  "/.github", "/.mypy_cache", "/.pytest_cache", "/.githooks",
  "sonar-project.properties"
]

[tool.hatch.build.targets.wheel]
packages = ["src/dtscalibration"]

[tool.hatch.publish.index]
disable = true  # Requires confirmation when publishing to pypi.

[project]
name = "dtscalibration"
version = "3.0.0"
description = "Load Distributed Temperature Sensing (DTS) files, calibrate the temperature and estimate its uncertainty."
readme = "README.rst"
license = "BSD-3-Clause"
requires-python = ">=3.9, <3.12"
authors = [
  {name = "Bas des Tombe, Bart Schilperoort"},
]
maintainers = [
  {name = "Bas des Tombe, Bart Schilperoort"},
]
keywords = [
  "DTS",
  "distributed temperature sensing",
  "calibration",
  "confidence intervals",
]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: BSD License",
    "Operating System :: Unix",
    "Operating System :: POSIX",
    "Operating System :: Microsoft :: Windows",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Utilities",
]
dependencies = [
    "numpy>=1.22.4",        # >= 1.22 for full quantile method support in xarray
    "dask",
    "pandas",
    "xarray[parallel]",     # numbagg (llvmlite) is a pain to install with pip
    "bottleneck",           # speeds up Xarray
    "flox",                 # speeds up Xarray
    "pyyaml>=6.0.1",
    "xmltodict",
    "scipy",
    "statsmodels",
    "matplotlib",
    "netCDF4>=1.6.4",
    "nc-time-axis>=1.4.1"  # plot dependency of xarray
]

[project.optional-dependencies]
dev = [
    "hatch",
    "bump2version",
    "ruff",
    "isort",
    "black[jupyter]",   # for formatting
    "mypy",             # for type checking
    "types-PyYAML",     # for pyyaml types
    "types-xmltodict",  # for xmltodict types
    "pandas-stubs",     # for pandas types
    "pytest",
    "pytest-cov",       # for coverage
    "pytest-xdist",     # for parallel testing
    "jupyter",
    "nbformat",         # Needed to run the tests
]
docs = [  # Required for ReadTheDocs
    "IPython",
    "myst_parser",
    "matplotlib>=3.0.0",
    "sphinx",
    "sphinx_rtd_theme",
    "sphinx-autoapi",
    "sphinx-automodapi",
    "sphinx-autosummary-accessors",
    "coverage[toml]",
    "nbsphinx",
    "ipykernel",
    'recommonmark',
    'sphinx-automodapi',
    'pypandoc',
    'jupyter_client',
]

[tool.hatch.envs.default]
features = ["dev"]

[tool.hatch.envs.default.scripts]
lint = [
  "ruff check src/ tests/ examples/",
  "black --check docs/notebooks",
  "isort --check-only --diff src/ tests/ examples/",
  "mypy src/",
]
format = ["isort src/ tests/ examples/", "ruff check src/ tests/ examples/ --fix", "ruff format src/ tests/ examples/", "black docs/notebooks", "lint",]
test = ["pytest ./src/ ./tests/",]
fast-test = ["pytest ./tests/ -m \"not slow\"",]
coverage = [
  "pytest --cov --cov-report term --cov-report xml --cov-branch --junitxml=xunit-result.xml tests/",
]

[tool.hatch.envs.docs]
features = ["docs"]

[tool.hatch.envs.docs.scripts]
build = [
    "sphinx-build -E -c docs -b html docs dist/docs",    #"python docs/nb_examples_to_docs.py",
]

[tool.hatch.envs.matrix_test]
features = ["dev"]

[[tool.hatch.envs.matrix_test.matrix]]
python = ["3.9", "3.10", "3.11"]

[tool.hatch.envs.matrix_test.scripts]
test = ["pytest ./src/ ./tests/",]  # --doctest-modules

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
]

[tool.ruff]
select = [  # It would be nice to have the commented out checks working.
  "E",  # pycodestyle
  "F",  # pyflakes
  # "B",  # flake8-bugbear
  # "D",  # pydocstyle
  # "C90",  # mccabe complexity
  # "N",  # PEP8-naming
  "UP",  # pyupgrade (upgrade syntax to current syntax)
  "PLE",  # Pylint error https://github.com/charliermarsh/ruff#error-ple
  # "PLR",  # Pylint refactor (e.g. too-many-arguments)
  "PLW",  # Pylint warning (useless-else-on-loop)
  # "I",  # isort
  "SIM",  # flake8-simplify

]
extend-select = [
  "D401",  # First line should be in imperative mood
  "D400",  # First line should end in a period.
  "D404",  # First word of the docstring should not be "This"
  "TID252",  # No relative imports (not pep8 compliant)
]
ignore = [
  "PLR2004",  # magic value used in comparson
  "E501",  # Line too long (want to have fixed
]
# Allow autofix for all enabled rules (when `--fix`) is provided.
fixable = ["E", "F", "UP", "PLE", "D"]
unfixable = []
line-length = 88
exclude = ["docs", "build"]
# Allow unused variables when underscore-prefixed.
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"
# Minimum supported Python version
target-version = "py39"

[tool.ruff.per-file-ignores]
"tests/**" = ["D"]

[tool.ruff.pydocstyle]
convention = "google"

[tool.ruff.mccabe]
max-complexity = 10

[tool.ruff.lint.isort]
force-single-line = true

[tool.isort]
py_version=39
force_single_line = true
known_first_party = ["dtscalibration"]
skip = [".gitignore", ".tox", "docs", ".venv"]
src_paths = ["src", "tests"]
line_length = 120

[tool.black]
line-length = 88
target-version = ['py39', 'py310', 'py311']
extend-exclude = ".venv"

[tool.mypy]
ignore_missing_imports = true  # Preferably false, but matplotlib, scipy and statsmodels are missing typing stubs
python_version = "3.9"

[tool.coverage.run]
branch = true
source = ["src/dtscalibration"]
command_line = "-m pytest"

[tool.coverage.report]
exclude_lines = [
  "pragma: no cover",
  "@overload",
  "if TYPE_CHECKING:",
  "if typing.TYPE_CHECKING:"
]
